Separate file for command definition
Move the enumeration for command definition in a separate file
